Move some things around, new router mod
