+++ rust/sealevel/programs/hyperlane-sealevel-token-native/src/processor.rs
@@ -1,7 +1,7 @@
 //! TODO
 
 use borsh::BorshSerialize;
-use hyperlane_sealevel_connection_client::RemoteRouterConfig;
+use hyperlane_sealevel_connection_client::router::RemoteRouterConfig;
 use hyperlane_sealevel_message_recipient_interface::MessageRecipientInstruction;
 use hyperlane_sealevel_token_lib::{
     instruction::{Init, TransferFromRemote, TransferRemote},
